// src/app/types/weather.ts
export interface WeatherRoot {
    coord: Coord
    weather: Weather[]
    base: string
    main: WeatherMain
    visibility: number
    wind: Wind
    clouds: Clouds
    dt: number
    sys: Sys
    timezone: number
    id: number
    name: string
    cod: number
}

export interface Coord {
    lon: number
    lat: number
}

export interface Weather {
    id: number
    main: string
    description: string
    icon: string
}

export interface WeatherMain {
    temp: number
    feels_like: number
    temp_min: number
    temp_max: number
    pressure: number
    humidity: number
    sea_level: number
    grnd_level: number
}

export interface Wind {
    speed: number
    deg: number
    gust: number
}

export interface Clouds {
    all: number
}

export interface Sys {
    country: string
    sunrise: number
    sunset: number
}